+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ## HEAD (Unreleased)
 * Upgrade to v3.41.0 of the AWS Terraform Provider
 * Add support for `aws.resourceGroupsTaggingApi` getResources
+* Add Principal for AppSync in nodejs SDK
 
 ---